chore: remove routes
