+++ packages/core/src/Route/routesToJSON.ts
@@ -1,4 +1,4 @@
-import { lodash, routeToChunkName } from '@umijs/utils';
+import { lodash, winPath } from '@umijs/utils';
 import { IRoute } from './types';
 
 interface IOpts {
@@ -27,10 +27,21 @@ export default function ({ routes, config, cwd }: IOpts) {
 
   function patchRoute(route: IRoute) {
     if (route.component && !isFunctionComponent(route.component)) {
-      const webpackChunkName = routeToChunkName({
-        route,
-        cwd,
-      });
+      const webpackChunkName = route.component
+        .replace(new RegExp(`^${lastSlash(winPath(cwd || '/'))}`), '')
+        .replace(/^.(\/|\\)/, '')
+        .replace(/(\/|\\)/g, '__')
+        .replace(/\.jsx?$/, '')
+        .replace(/\.tsx?$/, '')
+        .replace(/^src__/, '')
+        .replace(/\.\.__/g, '')
+        // 约定式路由的 [ 会导致 webpack 的 code splitting 失败
+        // ref: https://github.com/umijs/umi/issues/4155
+        .replace(/[\[\]]/g, '')
+        // 插件层的文件也可能是路由组件，比如 plugin-layout 插件
+        .replace(/^.umi-production__/, 't__')
+        .replace(/^pages__/, 'p__')
+        .replace(/^page__/, 'p__');
       route.component = [
         route.component,
         webpackChunkName,
@@ -82,4 +93,8 @@ export default function ({ routes, config, cwd }: IOpts) {
     })
     .replace(/\\r\\n/g, '\r\n')
     .replace(/\\n/g, '\r\n');
+}
+
+function lastSlash(str: string) {
+  return str[str.length - 1] === '/' ? str : `${str}/`;
 }
