+++ packages/utils/src/index.ts
@@ -50,7 +50,6 @@ export { traverse };
 export { signale };
 
 export * from './ssr';
-export * from './routes';
 export { default as compatESModuleRequire } from './compatESModuleRequire/compatESModuleRequire';
 export { default as mergeConfig } from './mergeConfig/mergeConfig';
 export { default as isLernaPackage } from './isLernaPackage/isLernaPackage';
